test(prd): cover queue pressure edge cases
